+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -8,8 +8,11 @@ import kotlin.Array
 import kotlin.Boolean
 import kotlin.CharSequence
 import kotlin.Comparable
+import kotlin.Double
 import kotlin.Int
+import kotlin.Long
 import kotlin.Nothing
+import kotlin.OverloadResolutionByLambdaReturnType
 import kotlin.Pair
 import kotlin.PublishedApi
 import kotlin.String
@@ -280,6 +283,24 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     public inline fun indexOfLast(predicate: (element: T) -> Boolean): Int {
         return values.indexOfLast(predicate)
     }
+
+    /**
+     * See [Array.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: T) -> Int): Int = values.sumOf(selector)
+
+    /**
+     * See [Array.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: T) -> Long): Long = values.sumOf(selector)
+
+    /**
+     * See [Array.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: T) -> Double): Double = values.sumOf(selector)
 
     /**
      * Returns an immutable array containing the first [n] elements.