+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableFloatArray.kt
@@ -7,9 +7,12 @@ import kotlin.Array
 import kotlin.Boolean
 import kotlin.CharSequence
 import kotlin.Comparable
+import kotlin.Double
 import kotlin.Float
 import kotlin.FloatArray
 import kotlin.Int
+import kotlin.Long
+import kotlin.OverloadResolutionByLambdaReturnType
 import kotlin.Pair
 import kotlin.PublishedApi
 import kotlin.String
@@ -280,6 +283,24 @@ public value class ImmutableFloatArray @PublishedApi internal constructor(
     public inline fun indexOfLast(predicate: (element: Float) -> Boolean): Int {
         return values.indexOfLast(predicate)
     }
+
+    /**
+     * See [FloatArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Float) -> Int): Int = values.sumOf(selector)
+
+    /**
+     * See [FloatArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Float) -> Long): Long = values.sumOf(selector)
+
+    /**
+     * See [FloatArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Float) -> Double): Double = values.sumOf(selector)
 
     /**
      * Returns an immutable array containing the first [n] elements.