+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -7,7 +7,10 @@ import kotlin.Array
 import kotlin.Boolean
 import kotlin.CharSequence
 import kotlin.Comparable
+import kotlin.Double
 import kotlin.Int
+import kotlin.Long
+import kotlin.OverloadResolutionByLambdaReturnType
 import kotlin.Pair
 import kotlin.PublishedApi
 import kotlin.Short
@@ -280,6 +283,24 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
     public inline fun indexOfLast(predicate: (element: Short) -> Boolean): Int {
         return values.indexOfLast(predicate)
     }
+
+    /**
+     * See [ShortArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Short) -> Int): Int = values.sumOf(selector)
+
+    /**
+     * See [ShortArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Short) -> Long): Long = values.sumOf(selector)
+
+    /**
+     * See [ShortArray.sumOf]
+     */
+    @OverloadResolutionByLambdaReturnType
+    public inline fun sumOf(selector: (element: Short) -> Double): Double = values.sumOf(selector)
 
     /**
      * Returns an immutable array containing the first [n] elements.